+++ src/main.jsx
@@ -6,8 +6,6 @@ import Navbar from "./components/Navbar";
 import Home from "./pages/Home";
 import Explore from "./pages/Explore";
 import Error from "./pages/Error";
-import Login from "./pages/Login";
-import Register from "./pages/Register";
 import UpdateProfile from "./pages/UpdateProfile";
 import DashboardAdmin from "./pages/DashboardAdmin";
 import AllUserAdmin from "./pages/AllUserAdmin";
@@ -18,6 +16,7 @@ import CategoryAdmin from "./pages/CategoryAdmin";
 import UpdateProfileAdmin from "./pages/UpdateProfileAdmin";
 import Residencies from "./components/Residencies";
 import Register2 from "./pages/Register2";
+import Login2 from "./pages/Login2";
 
 const router = createBrowserRouter([
   {
@@ -37,22 +36,7 @@ const router = createBrowserRouter([
       </>
     ),
   },
-  {
-    path: "/login",
-    element: (
-      <>
-        <Login />
-      </>
-    ),
-  },
-  // {
-  //   path: "/register",
-  //   element: (
-  //     <>
-  //       <Register />
-  //     </>
-  //   ),
-  // },
+
   {
     path: "/update-profile",
     element: (
@@ -134,12 +118,16 @@ const router = createBrowserRouter([
       </>
     ),
   },
+  {
+    path: "/login2",
+    element: (
+      <>
+        <Login2 />
+      </>
+    ),
+  },
 ]);
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  // <React.StrictMode>
-  // <Home />
-  // </React.StrictMode>,
-
   <RouterProvider router={router} />
 );
